requirements.txt
cbbd==1.23.0
pandas==2.2.3
pyarrow==15.0.2
python-dotenv==1.0.1
